Move LoadingContext to common/contexts/, formatters to common/utils/ and fetchWithAuth to config/auth/
