+++ client/src/pages/core/articulo/components/detail/index.jsx
@@ -27,8 +27,8 @@ import { useParams } from 'react-router-dom';
 import checkPermissions from '../../../../../config/auth/checkPermissions';
 import { API } from '../../../../../App';
 import SnackbarAlert from '../../../../../common/components/SnackbarAlert';
-import fetchWithAuth from '../../../../../utils/fetchWithAuth';
-import { useLoading } from '../../../../../utils/loadingContext';
+import fetchWithAuth from '../../../../../config/auth/fetchWithAuth';
+import { useLoading } from '../../../../../common/contexts/LoadingContext';
 
 
 export default function ArticuloDetail({ permissions }) {
