+++ client/src/pages/core/cliente/components/form/index.jsx
@@ -28,9 +28,9 @@ import dayjs from 'dayjs';
 import SimpleTabPanel from '../../../../../common/components/SimpleTabPanel';
 import { API } from '../../../../../App';
 import TributoDataGrid from '../../../../../common/components/TributoDataGrid';
-import fetchWithAuth from '../../../../../utils/fetchWithAuth';
+import fetchWithAuth from '../../../../../config/auth/fetchWithAuth';
 import SnackbarAlert from '../../../../../common/components/SnackbarAlert';
-import { useLoading } from '../../../../../utils/loadingContext';
+import { useLoading } from '../../../../../common/contexts/LoadingContext';
 import checkPermissions from '../../../../../config/auth/checkPermissions';
 import PageTitle from '../../../../../common/components/PageTitle';
