+++ client/src/pages/core/comercio/components/form/index.jsx
@@ -21,10 +21,10 @@ import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import dayjs from 'dayjs';
 import SaveIcon from '@mui/icons-material/Save';
 import { API } from '../../../../../App';
-import fetchWithAuth from '../../../../../utils/fetchWithAuth';
+import fetchWithAuth from '../../../../../config/auth/fetchWithAuth';
 import SnackbarAlert from '../../../../../common/components/SnackbarAlert';
 import SimpleTabPanel from '../../../../../common/components/SimpleTabPanel';
-import { useLoading } from '../../../../../utils/loadingContext';
+import { useLoading } from '../../../../../common/contexts/LoadingContext';
 import checkPermissions from '../../../../../config/auth/checkPermissions';
 import PageTitle from '../../../../../common/components/PageTitle';
