+++ client/src/pages/core/movimiento_stock/components/detail/index.jsx
@@ -21,8 +21,8 @@ import Edit from '@mui/icons-material/Edit';
 import dayjs from 'dayjs';
 import 'dayjs/locale/es';
 import SnackbarAlert from '../../../../../common/components/SnackbarAlert';
-import fetchWithAuth from '../../../../../utils/fetchWithAuth';
-import { useLoading } from '../../../../../utils/loadingContext';
+import fetchWithAuth from '../../../../../config/auth/fetchWithAuth';
+import { useLoading } from '../../../../../common/contexts/LoadingContext';
 import { API } from '../../../../../App';
 import checkPermissions from '../../../../../config/auth/checkPermissions';
 import PageTitle from '../../../../../common/components/PageTitle';
