+++ client/src/pages/core/venta/components/orden/list/index.jsx
@@ -14,8 +14,8 @@ import { Button, Box, Chip } from "@mui/material";
 import { Add, Visibility, Edit, Done, Block, Info, Delete } from '@mui/icons-material';
 import { useConfirm } from 'material-ui-confirm';
 import { API } from '../../../../../../App';
-import { useLoading } from '../../../../../../utils/loadingContext';
-import fetchWithAuth from '../../../../../../utils/fetchWithAuth';
+import { useLoading } from '../../../../../../common/contexts/LoadingContext';
+import fetchWithAuth from '../../../../../../config/auth/fetchWithAuth';
 import SnackbarAlert from '../../../../../../common/components/SnackbarAlert';
 import PageTitle from '../../../../../../common/components/PageTitle';
 import checkAuth from '../../../../../../config/auth/checkAuth';
